cleanup: drop unused vars, set references for object const parameters, optimisation, fix warnings, fix merge bugs, fix asan errors and more pvs studio fixes and more garbage again fix clang warns

+++ vita3k/ngs/src/modules/atrac9.cpp
@@ -139,7 +139,7 @@ bool Atrac9Module::decode_more_data(KernelState &kern, const MemState &mem, cons
             return true;
         }
         // make the byte position negative, will be positive at the end
-        state->current_byte_position_in_buffer = -old_size;
+        state->current_byte_position_in_buffer = -(int32_t)old_size;
         input = temp_buffer.data();
     }
